Updated canvas js
Fixed the angle indicator being offset and added speed. Added a setting in code called "anglebarheight" which is the height in pixels of each bar. This can be changed and the pitch indicator will change dynamically to the new height of the indicators. Added 3 variables called "planePitch", "planeRoll", "planeSpeed" which is the variables that you can change to change the indicators of the plane.